Fix for #42

+++ src/Rutorika/Sortable/ToSortedManyTrait.php
@@ -145,7 +145,9 @@ trait ToSortedManyTrait
      */
     public function sync($ids, $detaching = true)
     {
-        $this->detach();
+        if($detaching) {
+            $this->detach();
+        }
 
         return parent::sync($ids, $detaching);
     }
